Cygwin: store tls pointer in win32 tls
Use WIN32 TLS instead of stack-relative pointers for TLS. This allows windows
fibers and boost coroutines to be used without crashing whenever a syscall is
made from a fiber.

NOTE: This should be submitted upstream, but there is a blocker first. In the
main branch there are conflicts with new ARM64 support. In addition to fixing
the conflicts, we should actually test whether ARM64 builds work too, first.