Cygwin: only use native symlinks when they preserve the target
We want readlink() to return the original target string as much as
possible. Currently it'll be changed in some common cases:

x/ -> x
x/../x -> x

This doesn't normally break the links, but it means the link will change
when archived. With Nix, this can change the hash of a NAR packed tree.